@@ -22,12 +22,12 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
-import javax.persistence.EntityManager;
-import javax.persistence.NoResultException;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.NoResultException;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Root;
 import lombok.extern.slf4j.Slf4j;
 import org.keycloak.component.ComponentModel;
 import org.keycloak.models.KeycloakSession;
@@ -62,7 +62,7 @@ public class ScimResourceTypeService extends AbstractService {
     Optional<ScimResourceTypeEntity> optionalResourceTypeEntity =
         getResourceTypeEntityByName(resourceType.getName());
     if (!optionalResourceTypeEntity.isPresent()) {
-      log.info(
+      log.debug(
           "no database entry found for resource type {}. Entry will be created",
           resourceType.getName());
       return createNewResourceTypeEntry(resourceType, realmModel);
